Moving the dapr component checking logic to the handler
