+++ test/step/deployerrorexecutor.go
@@ -7,9 +7,11 @@ package step
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
+	"reflect"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -46,17 +48,21 @@ func (d *DeployErrorExecutor) Execute(ctx context.Context, t *testing.T, options
 
 	templateFilePath := filepath.Join(cwd, d.Template)
 	t.Logf("deploying %s from file %s", d.Description, d.Template)
-
 	cli := radcli.NewCLI(t, options.ConfigFilePath)
 	err = cli.Deploy(ctx, templateFilePath, d.Parameters...)
-	require.NotNil(t, err)
 	require.Error(t, err, "deployment %s succeeded when it should have failed", d.Description)
 
-	// TODO: Need to find out how to unwrap fmt.wrapError all the way up the stack
-	// so that it can be casted to other error types. Due to this issue, the error
-	// cannot be casted to conv.NewClientErrResourceConflict in the deployment processor.
-	if cliErr, ok := err.(*radcli.CliError); ok {
+	var cliErr *radcli.CliError
+	switch {
+	case errors.As(err, &cliErr):
+		t.Logf("error is a Rad CLI error")
 		require.Equal(t, d.ExpectedErrorCode, cliErr.Code)
+	default:
+		// t.Logf("error is not a Rad CLI error")
+		// t.Logf("error type is %q", reflect.TypeOf(err))
+		unwrappedError := errors.Unwrap(err)
+		t.Logf("error is not a Rad CLI error")
+		t.Logf("error type is %q", reflect.TypeOf(unwrappedError))
 	}
 
 	t.Logf("finished deploying %s from file %s", d.Description, d.Template)
